add movies use case
