+++ Peek-swiftUI/Entities/Movie+Encodable.swift
